+++ TamagotchiProject/AppDelegate.swift
@@ -11,6 +11,8 @@ import UIKit
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        
+//        UNUserNotificationCenter.current().delegate = self
         // Override point for customization after application launch.
         return true
     }
