+++ TamagotchiProject/setting/defaultUISetting.swift
